fix session reset on vercel

+++ settings.py
@@ -62,7 +62,7 @@ SOURCE_DB = BASE_DIR / 'db.sqlite3'
 if os.environ.get('VERCEL') == '1':
     TMP_DB = Path('/tmp/db.sqlite3')
 
-    if SOURCE_DB.exists():
+    if SOURCE_DB.exists() and not TMP_DB.exists():
         shutil.copyfile(SOURCE_DB, TMP_DB)
 
     DB_NAME = TMP_DB
